requirements.txt
discord.py == 2.0.1
catapi.py
pygelbooru
# pyspark
plotly
kaleido
pandas
toml
peewee
recordclass
emoji >= 2.0.0
suffix-trees == 0.3.0
yt-dlp
# bottomify
pyfiglet
cowsay
httpx
aiohttp
ics
cachetools
asyncache
yarl
more-itertools
python-dateutil
